Using create-react-app for example
