+++ .babelrc
@@ -4,7 +4,6 @@
     "@babel/preset-react"
   ],
   "plugins": [
-    "@babel/plugin-proposal-class-properties",
-    "react-hot-loader/babel"
+    "@babel/plugin-proposal-class-properties"
   ]
 }
