--- example/index.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0, user-scalable=no">
-    <title>React CSV Downloader</title>
-  </head>
-  <body>
-    <div id="app"></div>
-    <script src="http://localhost:8080/webpack-dev-server.js"></script>
-    <script src="./assets/js/bundle.js"></script>
-  <body>
-</html>
