+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "mocha --opts ./mocha.opts \"src/**/*.spec.js\"",
     "lint": "eslint .",
-    "start": "webpack-dev-server --progress --colors --hot --inline",
+    "start": "npm start --prefix example",
     "build": "babel src -d dist",
     "prepare": "npm run build"
   },
@@ -29,6 +29,7 @@
     "url": "https://github.com/dolezel/react-csv-downloader/issues"
   },
   "homepage": "https://github.com/dolezel/react-csv-downloader",
+  "dependencies": {},
   "devDependencies": {
     "@babel/cli": "7.6.2",
     "@babel/core": "7.6.2",
@@ -37,7 +38,6 @@
     "@babel/preset-react": "7.0.0",
     "@babel/register": "7.6.2",
     "babel-eslint": "10.0.3",
-    "babel-loader": "8.0.6",
     "chai": "4.2.0",
     "eslint": "6.5.0",
     "eslint-config-airbnb": "18.0.1",
@@ -47,12 +47,7 @@
     "mocha": "6.2.1",
     "prop-types": "15.7.2",
     "react": "16.10.1",
-    "react-dom": "16.10.1",
-    "react-hot-loader": "4.12.14",
-    "sinon-chai": "3.3.0",
-    "webpack": "4.41.0",
-    "webpack-cli": "3.3.9",
-    "webpack-dev-server": "3.8.1"
+    "sinon-chai": "3.3.0"
   },
   "peerDependencies": {
     "prop-types": "^15",
