--- webpack.config.js
@@ -1,28 +1,0 @@
-const path = require('path');
-
-module.exports = {
-  mode: 'development',
-  entry: {
-    'js/bundle': './example/index.js',
-  },
-  output: {
-    path: path.resolve(__dirname, 'example/assets'),
-    publicPath: '/assets/',
-    filename: '[name].js',
-  },
-  devServer: {
-    contentBase: './example',
-    inline: true,
-  },
-  module: {
-    rules: [
-      {
-        test: /\.js(x?)$/,
-        exclude: /(node_modules|bower_components)/,
-        use: {
-          loader: 'babel-loader',
-        },
-      },
-    ],
-  },
-};
